+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
+### Added
+* Ozon Content: Ondersteuning `<NieuweTekst>`, `<VerwijderdeTekst>`, `<table wijzigactie="verwijder">` & `<table wijzigactie="voegtoe">` ([#2060](https://github.com/dso-toolkit/dso-toolkit/issues/2060))
+
 ### Changed
 * **BREAKING**: Winkelwagentje open en dicht net zoals accordion. Gebruik icoon `chevron-right` ([#1984](https://github.com/dso-toolkit/dso-toolkit/issues/1984))